Move route constraints to /app/lib/route_constraints
