+++ app/lib/route_constraints/register_api_constraint.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module RouteConstraints
+  class RegisterApiConstraint
+    def self.matches?(request)
+      request[:api_version] == "v0.1"
+    end
+  end
+end
