+++ app/lib/route_constraints/system_admin_constraint.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module RouteConstraints
+  class SystemAdminConstraint
+    def matches?(request)
+      signin_user = OtpSignInUser.load_from_session(request.session) ||
+        DfESignInUser.load_from_session(request.session)
+      signin_user&.system_admin?
+    end
+  end
+end
