+++ config/routes/api_routes.rb
@@ -10,7 +10,7 @@ module ApiRoutes
       end
 
       namespace :api, path: "api/:api_version", api_version: /v[.0-9]+/ do
-        resource :info, only: :show, controller: "info", constraints: ValidRegisterApiRoute
+        resource :info, only: :show, controller: "info", constraints: RouteConstraints::RegisterApiConstraint
         match "*url" => "base#not_found", via: :all
       end
     end
